Revert "fix: ads on native (#1623)"
This reverts commit 896dc9617a9b0f879036e7fdd06613aae7725f32.
