+++ packages/ad/__tests__/web/utils.web.test.js
@@ -1,6 +1,6 @@
-import sharedAdInit from "../utils/ad-init.web";
+import sharedAdInit from "../utils/ad-init";
 import sharedAdInitGPT from "../utils/ad-init-gpt";
-import sharedAdInitPrebid from "../utils/ad-init-prebid.web";
+import sharedAdInitPrebid from "../utils/ad-init-prebid";
 import sharedAdInitUtils from "../utils/ad-init-utils";
 import sharedGenerateConfig from "../utils/generate-config.shared";
 import sharedPrebidConfig from "../utils/prebid-config";
